add support for interaction_code on redirect callback
utils to detect interaction_required error
