+++ test/e2e/package.json
@@ -23,9 +23,7 @@
     "@wdio/sauce-service": "^6.6.0",
     "@wdio/selenium-standalone-service": "^6.6.5",
     "@wdio/spec-reporter": "^6.6.0",
-    "chromedriver": "^87.0.0",
     "wait-on": "^3.3.0",
-    "wdio-chromedriver-service": "^6.0.4",
     "webpack": "^3.0.0"
   }
 }
